requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.5
django-allauth==0.44.0
django-bootstrap4==2.3.1
django-guardian==2.3.0
django-simpleui==2021.1.1
idna==2.10
oauthlib==3.1.0
pycparser==2.20
PyJWT==2.0.0
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
mysqlclient==2.0.1
gunicorn==19.9.0
django-widget-tweaks==1.4.8
django-prometheus==2.1.0
pandas==1.2.2
django-ckeditor==6.0.0
django-tinymce==3.3.0
openpyxl==3.0.7
Jinja2==3.0.1
jinja2-time==0.2.0
selenium==3.14.1
django_apscheduler==0.6.0
jieba==0.41
wordcloud==1.8.2.2
APScheduler==3.9.1
lxml==4.6.3
Pillow==8.4.0
numpy==1.20.3